pollapi/requirements.txt
Package             Version
------------------- -------------------
asgiref             3.3.1
certifi             2020.12.5
chardet             4.0.0
coreapi             2.3.3
coreschema          0.0.4
Django              3.1.4
django-filter       2.4.0
django-rest-swagger 2.2.0
djangorestframework 3.12.2
idna                2.10
importlib-metadata  3.3.0
itypes              1.2.0
Jinja2              2.11.2
Markdown            3.3.3
MarkupSafe          1.1.1
openapi-codec       1.3.2
pip                 20.3.3
pytz                2020.5
requests            2.25.1
restless            2.1.1
setuptools          51.0.0.post20201207
simplejson          3.17.2
six                 1.15.0
sqlparse            0.4.1
typing-extensions   3.7.4.3
uritemplate         3.0.1
urllib3             1.26.2
wheel               0.36.2
zipp                3.4.0